remove default templates
